// @types/generated/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IAboutPageConfigFields {
  /** Header */
  header: string;

  /** Headshot */
  headshot: Asset;

  /** Description */
  description: Document;

  /** Name */
  name: "About Page Config";
}

export interface IAboutPageConfig extends Entry<IAboutPageConfigFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "aboutPageConfig";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IBookFields {
  /** Cover */
  cover: Asset;

  /** Title */
  title: string;

  /** Author(s) */
  authors: string[];

  /** Tag Line */
  tagLine?: string | undefined;

  /** Description */
  description?: Document | undefined;

  /** Buy Links */
  buyLinks?: IBuyLink[] | undefined;
}

export interface IBook extends Entry<IBookFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "book";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IBuyLinkFields {
  /** Platform */
  platform: string;

  /** Link */
  link: string;
}

export interface IBuyLink extends Entry<IBuyLinkFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "buyLink";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ILandingPageConfigFields {
  /** Call To Action */
  callToAction: string;

  /** Carousel Books */
  carouselBooks: IBook[];

  /** Brief Bio */
  briefBio?: Document | undefined;

  /** Newsletter Signup */
  newsletterSignup: boolean;

  /** Name */
  name: "Landing Page Config";

  /** Call To Action Background */
  callToActionBackground: Asset;
}

export interface ILandingPageConfig extends Entry<ILandingPageConfigFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "landingPageConfig";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISeriesFields {
  /** Name */
  name: string;

  /** Tag Line */
  tagLine?: string | undefined;

  /** Description */
  description?: Document | undefined;

  /** Books */
  books: IBook[];

  /** Splash Art */
  splashArt?: Asset | undefined;

  /** Buy Links */
  buyLinks?: IBuyLink[] | undefined;
}

export interface ISeries extends Entry<ISeriesFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "series";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE =
  | "aboutPageConfig"
  | "book"
  | "buyLink"
  | "landingPageConfig"
  | "series";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";